safespace - version 0.8 - alpha
Added more icons for subsequent site visuals.
Updated the page css to look cleaner, center and resize site elements properly, on a variety of devices (PC, mobile).
Built the /home page for the site, adding on to previous work in progress and finishing the html to add 3 core features: search, creating posts, list of posts. Includes html, css, js, and python code for session handling and database interaction.
Added placeholders with documentative text for the /messages and /groups pages. Used the common format to create a placeholder for /profile as well, but with added python-database features to allow uploading of a profile picture.
Added crsf vulnerability protection.
Updated README progress and clarity.
App tested to work with the installation instructions.